refactor: clipboard service
